Provide a value to RealMemory in slurm.conf
This makes it possible to select nodes based on the available memory
by using the --mem parameter when submitting a job.
Since memory is not a consumable resource in the current config file
template, memory usage is not tracked. Slurm only checks that the value
passed to --mem is lower than the value of RealMemory from the config
file.