Increase the wait time for killing jobs in slurm
When executing scancel, jobs are now given one minute to respond to a
SIGTERM and two minutes to respond to a SIGKILL before being declared
unkillable. The defaults are respectively 30 seconds and one minute.
When slurm declares a job unkillable, it brings the node down until
manual maintenance is performed.