Fix a boot order problem with the docker socket
By default, the docker socket service starts up along with other sockets
managed by systemd, which might take place before NSS. As a result,
the the lookup for the docker group in IPA would fail and the docker
socket wouldn't start up on boot. This change instructs the docker
socket to start after NSS is online.