Make the snakemake profile a template
On slurm nodes, the slurm profile is installed. On all other compute
nodes, the local profile is installed. Note: in site.yml, the snakemake
role has been added to the managed_cluster_compute play, but it applies
to all compute nodes. The play will be retargeted in a separate change.