Move the custom ansible facts to their own role
There is only one custom fact at the moment, namely gpu['count'].
It is collected on cpu, gpu and kubernetes nodes. The immediate aim
is to have a mechanism for detecting the presence of gpus instead of
creating a separate inventory group for every task that needs a tweak on
GPU-equipped nodes. More generally, whenever we need another custom
fact, we can add it to the new role.